+++ src/app/layout.tsx
@@ -5,7 +5,7 @@ import { AppRouterCacheProvider } from "@mui/material-nextjs/v13-appRouter"
 import { ThemeProvider } from "@/app/theme-provider"
 import { HomeLayout } from "@/app/home-layout"
 
-export default function Layout({ children }: { children: ReactNode }) {
+const Layout = ({ children }: { children: ReactNode }) => {
   return (
     <html lang="en">
       <body style={{ margin: 0 }}>
@@ -18,3 +18,5 @@ export default function Layout({ children }: { children: ReactNode }) {
     </html>
   )
 }
+
+export default Layout
