+++ src/hooks/use-device-detect.ts
